requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# LLM
google-generativeai>=0.5.0

# UI - Chainlit (modern AI chat UI)
chainlit>=1.3.0

# Search Tools
tavily-python>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Memory & Vector Store
chromadb>=0.4.22
sentence-transformers>=2.5.0

# Email (optional)
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# PDF Generation
reportlab>=4.0.0
markdown>=3.5.0

# Optional: Streamlit (for comparison)
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
aiofiles>=23.0.0
